monitoring/zabbix/requirements.txt
py-zabbix
avisdk
